+++ client/src/components/LeftMenu/Channel.css
@@ -1,0 +1,23 @@
+.channel-item {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  color: #73777d;
+  font-size: 1.5rem;
+  margin: 5px 15px;
+  padding: 10px 15px;
+  border-radius: 6px;
+  background-color: transparent;
+}
+
+.channel-item.selected {
+  color: white;
+
+  background-color: #2c2d34;
+}
+
+.channel-item-name {
+  margin-left: 5px;
+  flex-grow: 1;
+  font: bold;
+}
