+++ client/src/components/LeftMenu/ChannelContainer.css
@@ -1,0 +1,34 @@
+.channel-section-header {
+  display: flex;
+  align-items: center;
+  margin-top: 30px;
+  margin-left: 5px;
+  margin-right: 15px;
+  color: #73777d;
+}
+.channel-section-header:hover {
+  color: white;
+}
+
+.channel-section-header .channel-section-arrow {
+  font-weight: lighter;
+  font-size: 20px;
+  transform: rotate(0.25turn);
+  transition: transform 0.3s ease-in-out;
+}
+
+.channel-section-header .channel-section-arrow.expanded {
+  transform: rotate(0turn);
+}
+
+.channel-section-header span {
+  flex-grow: 1;
+  font-weight: bold;
+  font-size: 20px;
+}
+
+.channel-section-header .channel-plus-btn {
+  cursor: pointer;
+  font-weight: bold;
+  font-size: 30px;
+}
