+++ client/src/components/LeftMenu/LeftMenu.css
@@ -2,89 +2,3 @@
   width: 300px;
   background-color: var(--leftmenu-bg-color);
 }
-
-.title-container {
-  cursor: pointer;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  height: 70px;
-  padding: 0 20px;
-  box-shadow: 0 1px 0 rgba(4, 4, 5, 0.2), 0 1.5px 0 rgba(6, 6, 7, 0.05),
-    0 2px 0 rgba(4, 4, 5, 0.05);
-}
-
-.title {
-  position: relative;
-  flex-grow: 1;
-  color: white;
-  font-size: 1.5rem;
-  font-weight: bold;
-}
-
-.title-icon {
-  font-size: 30px;
-  color: white;
-  width: 30px;
-  height: 30px;
-  margin-left: 10px;
-}
-
-.channel-section-header {
-  display: flex;
-  align-items: center;
-  margin-top: 30px;
-  margin-left: 5px;
-  margin-right: 15px;
-  color: #73777d;
-}
-.channel-section-header:hover {
-  color: white;
-}
-
-.channel-section-header .channel-section-arrow {
-  font-weight: lighter;
-  font-size: 20px;
-  transform: rotate(0.25turn);
-  transition: transform 0.3s ease-in-out;
-}
-
-.channel-section-header .channel-section-arrow.expanded {
-  transform: rotate(0turn);
-}
-
-.channel-section-header span {
-  flex-grow: 1;
-  font-weight: bold;
-  font-size: 20px;
-}
-
-.channel-section-header .channel-plus-btn {
-  cursor: pointer;
-  font-weight: bold;
-  font-size: 30px;
-}
-
-.channel-item {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  color: #73777d;
-  font-size: 1.5rem;
-  margin: 5px 15px;
-  padding: 10px 15px;
-  border-radius: 6px;
-  background-color: transparent;
-}
-
-.channel-item.selected {
-  color: white;
-
-  background-color: #2c2d34;
-}
-
-.channel-item-name {
-  margin-left: 5px;
-  flex-grow: 1;
-  font: bold;
-}
