+++ client/src/components/LeftMenu/LeftMenuTitleBar.css
@@ -1,0 +1,26 @@
+.title-container {
+  cursor: pointer;
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  height: 70px;
+  padding: 0 20px;
+  box-shadow: 0 1px 0 rgba(4, 4, 5, 0.2), 0 1.5px 0 rgba(6, 6, 7, 0.05),
+    0 2px 0 rgba(4, 4, 5, 0.05);
+}
+
+.title {
+  position: relative;
+  flex-grow: 1;
+  color: white;
+  font-size: 1.5rem;
+  font-weight: bold;
+}
+
+.title-icon {
+  font-size: 30px;
+  color: white;
+  width: 30px;
+  height: 30px;
+  margin-left: 10px;
+}
